requirements.txt
discord.py==1.5.1
bs4==0.0.1
requests==2.25.1
